feat:add hook confirm to delete

+++ components/data-table.tsx
@@ -26,6 +26,7 @@ import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
 import { useState } from 'react';
 import { Trash } from 'lucide-react';
+import { useConfirm } from '@/hooks/useConfirm';
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
@@ -45,6 +46,10 @@ export function DataTable<TData, TValue>({
   const [sorting, setSorting] = useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([]);
   const [rowSelection, setRowSelection] = useState({});
+  const [ConfirmDialog, confirm] = useConfirm(
+    'Are you sure?',
+    'You are about to perform a bulk delete',
+  );
 
   const table = useReactTable({
     data,
@@ -65,7 +70,8 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
-      <div className="flex items-center py-4">
+      <ConfirmDialog />
+      <div className="flex items-center justify-between py-4">
         <Input
           placeholder={`Filter ${filterKey}...`}
           value={(table.getColumn('name')?.getFilterValue() as string) ?? ''}
@@ -74,23 +80,26 @@ export function DataTable<TData, TValue>({
           }
           className="max-w-sm"
         />
+        {table.getFilteredSelectedRowModel().rows.length > 0 && (
+          <Button
+            className="gap-2 text-xs font-normal"
+            variant="outline"
+            size="sm"
+            disabled={disable}
+            onClick={async () => {
+              const ok = await confirm();
+              if (ok) {
+                // @ts-ignore
+                onDelete(table.getFilteredSelectedRowModel().rows);
+                table.resetRowSelection();
+              }
+            }}
+          >
+            <Trash className="size-3" />
+            Delete ({table.getFilteredSelectedRowModel().rows.length})
+          </Button>
+        )}
       </div>
-      {table.getFilteredSelectedRowModel().rows.length > 0 && (
-        <Button
-          className="mb-4 gap-2 text-xs font-normal"
-          variant="outline"
-          size="sm"
-          disabled={disable}
-          onClick={() => {
-            // @ts-ignore
-            onDelete(table.getFilteredSelectedRowModel().rows);
-            table.resetRowSelection();
-          }}
-        >
-          <Trash className="size-3" />
-          Delete ({table.getFilteredSelectedRowModel().rows.length})
-        </Button>
-      )}
       <div className="rounded-md border">
         <Table>
           <TableHeader>
